% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_genes}
\alias{filter_genes}
\title{Filtering function}
\usage{
filter_genes(x, num_reads = 5, num_samples = 5)
}
\arguments{
\item{x}{a}

\item{num_reads}{a}

\item{num_samples}{a}
}
\description{
This function filters genes.
}
